[libqofono] Added QOfonoConnectionContext::disconnect()
It may be convenient to synchronously deactivate the context, for example,
prior to modifying settings. Other SetProperty calls will fail while context
is still active.

Also, it doesn't make sense to check the active status in setAccessPointName
because:

a) Name isn't the only property that can't be changed while context is active
b) It could be that Active has already changed on the ofono side but the
   property change notification hasn't reached us yet.

Better let ofono check the value of the Active property and make appropriate
decision based on that.